Moved net lib tests into tests.rs
